auth/jwt-token-tester/requirements.txt
PyJWT
cryptography

